+++ src/esm/components/dropdowns/user-menu/index.js
@@ -176,7 +176,7 @@ export var UserMenu = /*#__PURE__*/function (_Component) {
         title: this.props.buttonTitle,
         notifications: this.props.buttonNotifications,
         badgeColor: this.props.badgeColor,
-        textClassName: "k-Header__userMenuText--withEllipsis"
+        textClassName: "k-UserMenu__text--withEllipsis"
       });
     }
   }, {